Convert from double to float

+++ color.cuh
@@ -11,7 +11,7 @@
 
 using color = vec3;
 
-inline double linear_to_gamma(double linear_component) {
+inline float linear_to_gamma(float linear_component) {
     return sqrt(linear_component);
 }
